Added commit form to problem page
Added commit form (message field and “commit” button) to the problem
page. This is so a user can make note of his/her commit before actually
pushing the updated problem to GitHub. I also fixed a portion of the
GitHub controller where updates to the watched repo variable were not
being rendered on the UI. This was fixed using additional $scope.$apply
blocks.